// src/app/components/AssessmentStage/ApproveGroupings/approve-groupings.interface.tsx
export interface IFinalizeStageProps {
  _id: string;
  sessions: IFinalizeSessionProps[];
}

export interface IFinalizeSessionProps {
  _id: string;
  stageId: string;
  playerAssessmentIds: string[];
  drillIds: string[];
  practicePlanIds: string[];
}

export interface IFinalizeSessionFields {
  _id: string;
  stageId: string;
  playerAssessments: IPlayerAssessment[];
  drills: IDrill[];
  practicePlans: IPracticePlan[];
}